requirements.txt
pyguetzli
azure-functions
